Moved the item middleware to spot_item.py
